Refactor shader classes into separate files and packages
