+++ src/main/java/blusunrize/immersiveengineering/api/shader/ShaderRegistry.java
@@ -11,7 +11,8 @@ package blusunrize.immersiveengineering.api.shader;
 import blusunrize.immersiveengineering.api.Lib;
 import blusunrize.immersiveengineering.api.crafting.IngredientStack;
 import blusunrize.immersiveengineering.api.shader.CapabilityShader.ShaderWrapper;
-import blusunrize.immersiveengineering.api.shader.ShaderCase.ShaderLayer;
+import blusunrize.immersiveengineering.api.shader.ShaderLayer;
+import blusunrize.immersiveengineering.api.shader.impl.*;
 import blusunrize.immersiveengineering.common.IERecipes;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
