+++ src/main/java/blusunrize/immersiveengineering/api/shader/impl/ShaderCaseShield.java
@@ -1,27 +1,29 @@
 /*
  * BluSunrize
- * Copyright (c) 2018
+ * Copyright (c) 2017
  *
  * This code is licensed under "Blu's License of Common Sense"
  * Details can be found in the license file in the root folder of this project
  */
 
-package blusunrize.immersiveengineering.api.shader;
+package blusunrize.immersiveengineering.api.shader.impl;
 
 import blusunrize.immersiveengineering.ImmersiveEngineering;
+import blusunrize.immersiveengineering.api.shader.ShaderCase;
+import blusunrize.immersiveengineering.api.shader.ShaderLayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 
 import java.util.Collection;
 
-public class ShaderCaseBanner extends ShaderCase
+public class ShaderCaseShield extends ShaderCase
 {
-	public ShaderCaseBanner(ShaderLayer... layers)
+	public ShaderCaseShield(ShaderLayer... layers)
 	{
 		super(layers);
 	}
 
-	public ShaderCaseBanner(Collection<ShaderLayer> layers)
+	public ShaderCaseShield(Collection<ShaderLayer> layers)
 	{
 		super(layers);
 	}
@@ -29,7 +31,7 @@ public class ShaderCaseBanner extends ShaderCase
 	@Override
 	public ResourceLocation getShaderType()
 	{
-		return new ResourceLocation(ImmersiveEngineering.MODID, "banner");
+		return new ResourceLocation(ImmersiveEngineering.MODID, "shield");
 	}
 
 	@Override
@@ -41,6 +43,8 @@ public class ShaderCaseBanner extends ShaderCase
 	@Override
 	public boolean renderModelPartForPass(ItemStack shader, ItemStack item, String modelPart, int pass)
 	{
+		if("flash".equals(modelPart)||"shock".equals(modelPart))
+			return pass==getLayers().length-1;
 		return true;
 	}
 }
